requirements.txt
pydantic>=2.0.0
chromadb>=0.4.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-anthropic>=0.1.0
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
pytest>=8.0.0
streamlit>=1.30.0
fpdf2>=2.7.0
